fix: Replace Lombok @Data annotations from entity classes with @Builder, @Getter, @Setter, and @ToString
Lombok's generated equals() and hashcode() methods rely on fields, which
may change during persistence; and the default implementation of the
methods compare objects rather then checking for if objects represent
the same record in the db. Therefore, manually overriding the methods is
a must.
@Builder automatically generates an implementation of the builder
pattern allowing for construction of  objects in steps. It handles
optional parameters well, improves readability, and simplifies complex
object construction.