feat(controller/service): Updated the updateProduct method to allow soft deleting of ProductVariants with constraints and QOL changes
Some QOL features were considered in the updateProduct method:
  - When adding a new ProductVariant, the available and active flag are
autofilled to true if those fields were not included in the update
request.
  - When updating an existing ProductVariant, all fields do not need to
included in the update request. Fields without values in the update
request will use the current field values instead.

You cannot set the active flag of a Product to false if none of its
variants do not have an active flag set to true.

DTOs were created to make updating products easier.

An attempt to be more specific with exception handling messages was
made.

TODO: Add Logging to GlobalExceptionHandler methods