feat: Add patch request to update Order fields and items
Enables updating of order details and items via a patch request.
Supports modifications to order status, special instructions, pickup
time, and items with their quantities and notes.

How the functionality will fit into the application will need to be
fleshed out as editing orders which affect price after a payment is made
is problematic.
  - Allow a certain timeframe for a customer to edit their order e.g. 24
hours
  - Allow minor edits such as special instructions up until the order
enters production
  - Allow modifications to orders affecting price if the order hasn't
been processed yet OR If the order is aleady in production, allow a
change request.