feat(DatePicker): Integrate DatePicker element with RHF so it can be validated on form submission; Create story for calendar to better assess UI design
The DatePicker can be validated via RHF however, the validation method
here is a placeholder. The ConfiguredPickupRules are needed to generate
the zod schema so that the rules can be used to define the validation
for pickup dates.
Some minor changes to colour desing of the DatePicker, but still not
finalised.