chore/fix(frontend/disclosure): Add disclosure component containing text area for special instructions to Checkout; Prevent focus styling from appearing on button click
The DisclosurePanel container is styled with overflow-hidden which was
causing issues. It meant that any children with a focus styling that
extended focusRing would have its ring cut off on the left and right
sides. To work around this, the width of the panel was increased so that
it would extend a certain width from both sides of its container thus
allowing room for the focus ring.

The disclosure for the special instructions is still not finished. It
needs to be added to the different layouts.