refactor(frontend/Checkbox): Remove scaling animation
The scaling animation was inconsistent across browsers - it would appear
jittery - so it was removed.
The focus visible styling also seemed to be inconsistent - sometimes not
appearing - which may have been a result of a styling conflict with a
motion component. The focus styling is applied to a vanilla html element
instead to resolve this.

- Add hover styling where an area around the Checkbox is highlighted via
background colour change. The ring-offset colour is changed when the
Checkbox is hovered and is focus-visible to maintain this styling change