feat(frontend/Checkbox): Add rightSlot prop to DisclosureRadioHeader
The right slot prop is used to display additional content for example,
it is used in PaymentForm to show valid issuing banks for card payments.

- Add animated list of issuing banks for card payments in
PaymentForm.tsx. They are located in the card DisclosureRadioHeader,
where inactive cards are greyed out
- Add logos to the rightSlot of relevant DisclosureRadios e.g. Apple,
Google, and PayPal
- Prevent card SVGs from animating in when page is first mounted
- Remove width and height attributes from SVG files to allow the SVGs to
scale properly
- Introduce styling changes to DisclosureRadioPanel to better visually
seperate the panel from the header