fix(frontend/Checkout): Fix interactions between update preference field and contact fields
If, on a newly mounted page, the update preference was chosen to be
email and then an email was provided afterwards, the field error on
update preference would not disappear. Therefore, a new case for
triggering validation - from the contact fields - for update preference
was to check if the update preference field was dirty.

- Add hasContactError as a prop to ControlledPhoneField.tsx to make it
re-render when errors.contact changes - it was previously not updating
its UI when errors.contact  had appeared from interacting with the email
field
- Add press animation to Checkbox to make the component more dynamic and
expressive