feat(frontend/Checkout): Create a reusable type-safe API client
Create a new apiClient.ts to centralise and simplify making API calls
with measures to ensure errors are handled and type-safety:
- Native fetch doesn't throw an error for 400 and 500 status codes, so
ApiError is manually thrown to trigger a failure which a try-catch block
can catch and handle.
- isBackendError is a type guard function that checks if an unknown
error object matches the shape of a bad response from the backend so
that it can be safely used.

apiClient.ts is used in CheckoutSessions.tsx to make a post request to
create a new order. The current configuration does not yet handle and
display the errors received from bad responses or handle successes.
However, the post request can be sent with the correct success response
and the correct error responses (both logged). There are some other
issues that need to be addressed:
- Cash is an unsupported payment method when creating an order - this
can be resolved by amending this or removing cash as a payment method.
- Name fields must be checked if they contain a whitespace and no
special characters to be considered valid.

Other changes include:
- Add .env file to create environment specific values such as the base
api url
- Fix typo for productVariantId