feat(backend): Allow multiple validation messages for a single field
Hibernate Validator's default fail-fast behaviour prevents multiple
constraint violations from being reported at the same time for a single
field. Disabling this behaviour globally was attempted but proved
unreliable.

This commit introduces a custom composite validator (@ValidCustomerName)
implemented in CustomerNameValidator.java. This validator performs both
pattern and length checks for name fields. If multiple constraint
are violated, then the corresponding error messages (provided by
MessageSource) are listed together.
This replaces the use of @Size and @Pattern annotations for name fields
in OrderRequestDto.java and CardDetailsRequestDto.java.

Unless there is a more resolution to the fail-fast behaviour, this will
need to be applied to other fields with more than 2 constraints e.g.
cardNumber and cvv.

ErrorReply.java has an updated type for validationMessages which is now
a Map<String, List<String>> which allows multiple constraint violation
messages to be listed for a violating field.