feat(backend/frontend): Make validationErrors a structured error object
Adding more detail to the validation errors provides better error
categorisation (Size, Pattern, etc.) and allows for better error
handling in the frontend as different UI elements can be conditionally
displayed based on error type.

Another change maded is the removal of CustomerNameValidation.java and
ValidCustomerName.java as they have become functionally redundant. The
compound validator was introduced as it was believed that
validationErrors was limited to a single constraint violation because of
a fail-fast validation behaviour. This was not the case -
validationErrors wasn't a list so only a single error was passed to the
field although others may have been present.

Additional changes:
- Update field names in OrderRequestDto to match the field names used in
the frontend form so that errors can easily be matched to the
corresponding violating field in the frontend. order.types.ts is updated
to match this change.