feat(fronted/checkout): Add context for sharing server-side errors across components
Adds ServerErrorContext and useServerErrors to manage and propagate
backend validation errors.
This allows server-generated errors to be displayed alongside their
corresponding fields and to persist until the next submission - since
the server-side validation logic is seperated from the client-side..
Currently, only the fullName field makes use of its server-generated
error for its isInvalid state and for its errorMessage. This needs to be
repeated for all field.

An additional change was to introduce a new order type ValidationError
so that the contents of each error for a given field can be accessed.