fix(frontend/backend): Make name fields more permissible
The previous iteration of validation was too restrictive considering the
wide breadth of different names. This commit updates the pattern for
validating name and other string fields (billing address) by using a
minimal disallowed list containing control characters which can be
harmful to the system. In the frontend, the name fields are normalised
by collapsing consecutive whitespaces and trimming.