feat(backend/card): Update the expiration field to a YearMonth object
This change simplifies the whole process of handling expiration.
Previously, the expiration was kept as a string at a DTO level and as 2
seperate strings - for year and month -  at the entity level. The move
to a YearMonth object removes the need for additional conditional
validation for expiration.
2 important changes were required to make this change:
1. The annotation @JsonFormat which is needed to configure how the
incoming object is deserialized and serialized i.e. we can specify how
Jackson should deserialize the incoming MM/YY string from the request
into the YearMonth object used by the DTO.
2. Adding YearMonthDateAttributeConverter, a JPA converter to handle
persisting YearMonth objects. It correctly maps the object to a standard
DATE column in SQL which ensures the data is queryable. Without this,
the YearMonth would be stored as a binary blob which is not queryable or
human-readable.