fix(backend/order): Update validation logic for card number
The card number in CardDetailsRequestDto.java now has matching
validation logic with the frontend - a combination of Luhn's algorithm
and card number length matching with the accepted bank (Visa, MC, and
Amex). A unit test class is created to test the card number validation
logic using mockito.

The function luhnAlgorithm in the frontend had a redundant line removed.