feat(frontend/checkout): Refine animation and styling for disclosure radio group
The previous animation revealed gaps between the disclosures when a new
disclosure opened due to the spring animation that was applied to the
layout changes. To avoid this, a negative margin is applied so that any
spring transition applied to the movement of the disclosure did not
expose the background behind it.

The biggest change is to the styling. The borders are removed in favour
of a background colour to create visually grouped disclosures. This is
because the borders were difficult to visually optimise as animating
from a border to no border would create unsmooth animation.

Another change, was to animate the border radius using motion as it is
easier to orchestrate and manage.