feat(frontend/checkout): Add styling and animation to server error messages
When the server error message first appears, a shaking animation plays
on a warning icon. If the form is submitted again and any server errors
are still present, then the animation repeats. When an error message is
resolved - and the form is submitted -  a simple fade and lift animation
plays to remove the message from the DOM.