fix(frontend/checkout): Prevent Payment DisclosureRadioGroup from animating its layout
The layout prop is applied to each DisclosureRadio which makes it so -
that with each re-render -  the components would check if their size or
position had changed and would animate any of those changes. However,
the intention of applying this prop was to animate the positions of the
DisclosureRadios whenever one of them opened, but not having them
animate their repositioning when external elements affected their
positioning. To prevent this, the prop layoutDependency is applied with
the expandedIndex being passed to it. This changes the behaviour of the
DisclosureRadios by having them animate their position only when the
expandedIndex is updated after a re-render.