feat(frontend/checkout): Add hover indication for payment methods
When the DisclosureRadioHeader of an unselected payment method is
hovered, a simple x-translation animation plays on the "radio button"
and the method name. It makes use of a motion prop called custom which
takes dynamic component data to be passed to the variant function. This
allows for different animation properties, and so the hover animation
can be disabled for a DisclosureRadio that isSelected.

The first attempt at trying to conditionally apply the hover animation
didn't work as the whileHover variant was being conditionally applied.
The problem with this is that the whileHover prop is often only
evaluated on the initial render. When the component is first loaded and
it is defaulted to be selected, the whileHover evaluates to undefined
and a hover listener is never attached to the specific component
instance (whileHover={isSelected ? undefined : "hover"}). This means
that the hover animation is disabled reguardless of its isSelected
state.