fix(frontend/checkout): Adjust layout of components for certain window widths
The useWindowDimensions hook is updated to make use of the
useLayoutEffect hook. This React hook allows you to run code after React
has calculated all the DOM changes BUT before the browser paints those
changes to the screen. Previously, the useEffect hook was used but that
only allows code to be ran after a new screen is painted. Therefore, it
is more optimal to use the useLayoutEffect hook for measuring an
element's size and position.

The available card providers SVGs in the card number field are removed
from the DOM between the window widths of 900px and 768px so that it
doesn't overlap with the field's text. The same SVG indicator is in the
DisclosureRadioHeader so information is not lost.

The cvv and expiration date fields layout now changes to form a single
column before the expiration date field label wraps to a new line which
made the fields misaligned.