fix(frontend/checkout): Improve morphing effect in DisclosureRadioGroup animation
Before, the disclosures were separated using a margin. Any margin
updates were not animated which wasn't effective in creating the
morphing effect. However, when margin is animated it leads to unwanted
layout shifts in content above the DisclosureRadioGroup. Therefore, the
margin animation was replaced in favour of a translation animation which
does not affect the layout of surrounding DOM elements.