fix(frontend/checkout): Add offset when scrolling to a DisclosureRadio
The addition of an offset - via scrollMarginTop - ensures that the
height of the sticky header in the checkout can be accounted for when
scrolling to the DisclosureRadio. The logic for scrolling is now handled
in DisclosureRadio rather than the panel since the scroll will be to the
DisclosureRadio. A new prop is added to disable the scrolling behaviour.