feat(frontend/checkout): Update scrolling logic for components to consider the sticky header
The style scroll-padding-top is applied to the document when Checkout is
first mounted, which is set to the height of the sticky header. This
allows the nearest scrolling block to consider elements hidden behind it
as "off-screen".

The Disclosure and DisclosureRadio components have updated scroll margin
offset values to give the components more breathing room when scrolled
to. They also have the scrolling block set to nearest as to not disrupt
the checkout flow. When the scrolling block was set to start - for the
payment options - the whole process of going through the different
options felt too all over the place.