feat(frontend/order-confirmation): Create an OrderConfirmation page
This new Order Confirmation page displays some information from the
response received from a successful order creation. This content is
passed from the onSubmit function in the checkout page using React
Router's location state. The content on the page is limited at the
moment, but the all the data needed to populate the page is ready and
available.

The OrderConfirmation page is only navigable from the checkout page upon
a successful order placement. If the OrderConfirmation endpoint is
accessed outside of these condition they are redirected to the homepage.