feat(backend/checkout): Add pickup slot for CustomerOrders
Adding pickup slot to CustomerOrders makes the web application feel more
realistic as it mimics real ordering systems and shows that business
logic is being considered. It is made a mandatory field.

The pickup slots are represented by an enum where each constant in the
enum (like SLOT_1) is associated with a TimeRange object that defines
its start and end time. Further development needs to be made to
communicate this information with the client. Currently, the client is
unaware of the pickup slots but it passes a hard coded slot value in the
order placement request to maintain a working system. This will be user
provided in the future.