feat(frontend/checkout): Add layout styling to the pickupDate and pickupSlot fields
A div element now wraps the pickupDate and pickupSlot fields with a grid
layout of 2 columns. On a lg breakpoint, the grid layout is 2 columns  -
so the fields are side-by-side - to make the most of the horizontal
space. The only other time the grid is 2 columns is between sm and md
breakpoints.