feat(frontend/checkout): Update disabled state styling for text and select elements
The disabled state stylings are updated for fieldBorderStyles,
borderedSelectButtonStyles, and selectButtonStyles to better distinguish
the disabled state from their normal state stylings which were using
darker colours. A lighter colour is used - assigned to a tailwind
variable disabled-text - instead. The chevron used in the Select
component is also updated to use this colour when the component
isDisabled.