feat(frontend/datePicker): Update disabled styling for datePicker component
The disabled styling for the datePicker component now matches the
standard disabled field styling introduced in the previous commit, using
lighter colours. The opacity of the CalendarDays icon is decreased
slightly to better visually match the text colour. Although, they are
both using the same colour, the icon appears darker due to the stroke
thickness if at an opacity of 1.