feat(frontend/PhoneField): Update disabled styling for the PhoneField component
The PhoneField's disabled state styling now matches the standard across
other fields. In addition to this, the disabled state is now passed to
the Select component -  within the PhoneField component - to ensure all
interaction with all parts of the field are not permitted. The custom
select value has decreased opacity when disabled to enforce that the
component cannot be interacted with.