feat(frontend/utils): Add utility function for creating a debounce function and add a hook for creating a debounce function for triggering validation on checkout form fields
The useDebouncedFormTrigger hook is created as a solution to the lag in
typing in the email and phone number fields in the checkout form. This
was caused from the fields triggering validation on multiple fields
whenever the field value changed. The debounce delays the triggered
validations once the field is unchanged after a set delay - defaulted at
500ms for the hook - which offsets the computational load when typing.

An important note on the hook's behaviour is that the returned function
is memoized, so the same function instance persists across renders.
Since it's a single instance, multiple calls within the debounce delay
will reset the timer and only the arguments from the final call will be
executed which overrides all previous/pending calls. Therefore, it is
necessary to batch all field names into a singlecall instead of calling
the function multiple times with different arguments.

The hook is built on the utility function debounce.