feat(frontend/checkout): Add debounced validation to the email field
The email field is debounced using the useDebouncedFormTrigger hook. It
is important to bypass the default validation behaviour defined by RHF
to debounce the validation for the field itself as well as the contact
and updatePreference fields. Therefore, it was important to call
setValue - in the onChange prop - rather than the onChange event to have
the option to opt out of the default validation and opt into the
debounced validation. If onChange is called instead, the result of
interacting with the field would stay the same however, the process of
interacting with it would differ. For example, if the field was cleared
there would be a brief moment where the field would show no errors but
then the debounce function would be called and the contact error would
appear. This is because the onChange event would validate the field
immediately.

The onBlur handler is updated to call the same validation triggers as the onChange handler (with no
debounce) as a safety measure.

It is important to take note of the usage of the debounce hook. The
field names are batched into a single call to prevent the issue of
previous/pending calls - in the onChange handler - being
cancelled/overridden if the calls were made separately for each field.