feat(frontend/checkout): Add debounce validation to the phone number field
Moving to a debounced validation solution for the phone number field was
necessary to solve the issue with typing lag. Previously, the validation
for the updatePreference and contact fields, and itself was triggered
with every keystroke in the onPhoneNumberChange handler. Debouncing the
trigger function delays the validation until after the user stops typing
for a set period of time. This allows the user to type without
experiencing any lag due to validation computations. The
onCountryCodeChange handler remains the same, using the onChange
function, which automatically validates the field. The onBlur prop now
calls the RHF onBlur function and the same validation triggers in the
onPhoneNumberChange handler as a safety measure.

Another change made was to make the ControlledPhoneField component
"smarter" by removing any props being passed to it from the CustomerForm
component and replacing them with its own internal self-contained logic.
Since the component is called within the RHF form, it has access to the
useFormContext hook and so has access to all the data it needs to
function.

The debounce delay in useDebounceFormTrigger is shortened to make
validation feel more responsive.