feat(frontend/checkout): Match validation behaviour for the email and phone number fields with the rest of the form fields before the first for submission
The debounce validation for the email and phone number fields is only
triggered after the first form submission or if either of the 2 fields
are touched. This is so that its behaviour matches the default
"onTouched" mode set for the form. The reason for the mismatch is due to
the fact that the implementation of the debounced validation circumvents
the form's default validation behaviour so it needs to be adjusted to
match.

For this to work the "shouldTouch" option for the setValue functions, in
the onChange handlers, were removed. This would cause the form to treat
the field as touched when its value first changed -  which is not the
intention - and the debounce validation would trigger immediately before
the field was blurred.