feat(frontend/OrderConfirmation): Add styled table for order items
There a few styling and content changes to the OrderConfirmation page.
The biggest one being the addition of a table listing the ordered items;
their images (to be implemented later), their quantities, their totals
and a total of all items.

Other content changes include the addition of a title with the order
number; a message describing future steps after the order is confirmed;
and a pickup map. These changes are not finalised but give a better idea
of what content is going to be included and how it is going to mapped
out on the page.