feat(frontend/OrderConfirmation): Format spacing of table cells
The cells of the table now have uniform spacing between them. To achieve
this, all the order item cells are targeted to have bottom padding. The
first order item cell has top padding, in addition to the bottom
padding, to maintain the visual spacing between the table content above
it i.e. the column header. Margins do not exist on table cells in the
standard CSS box model so padding is used instead. The cellSyles
tailwind variant is removed as it felt redundant.