feat(backend/CustomerOrder): Return pickup slot label in OrderSuccessDto rather than the value
The pickup slot is now returned in the OrderSuccessDto instead of its
Enum value. This is because the label is needed to display in the
client. This is achieved by labelling the getLabel() function - in the
PickupSlot Enum - with @JsonValue which tells Jackson to use the result
of the method when converting the Enum to JSON.