feat(frontend/PickupMap): Add animated icons for indicating map actions expand and close
- Add an animated Expand icon component which can either be controlled
or uncontrolled. It will play its animation when hovered in its
uncontrolled state.
- Add an animated Shrink icon component which can either be controlled
or uncontrolled and plays a different animation depending on the values
of its 'isHovered' and 'isFocusedWithin' props.
- Add a 'peek' animation when the Shrink icon has 'isFocusedWithin' set
to true to better communicate that the user has landed on the icon
without permanently changing the icon's shape which can look 'stuck'.
- Add the Expand icon to the inline map's expand button.
- Add the Shrink icon to the expanded map's close button.

Todo: Implement the 'peek' animation for the Expand icon.